# swagger_output.json
{
  "swagger": "2.0",
  "consume": [
    "application/json"
  ],
  "produce": [
    "application/json"
  ],
  "info": {
    "title": "10DaysProject API",
    "description": "This is 10DaysProject API for our website Exam-Resource sharing Center.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/article": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "consume": [
          "application/json"
        ],
        "produce": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Article ObjectId",
            "required": true,
            "type": "string",
            "description": "ObjectId of an article auto-generated by Mongodb"
          },
          {
            "in": "body",
            "description": "The new Article to create",
            "schema": {
              "$ref": "#/definitions/Article"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/article/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Article": {
      "required": [
        "title",
        "link"
      ],
      "properties": {
        "title": {
          "type": "string",
          "required": true
      },
      "description": {
          "type": "string"
      },
      "link": {
          "type": "string"
      },
      "viewCount": {
          "type": "integer",
          "default": 0
      } 
      }
    }
  }
}